+++ apps/switch/main.py
@@ -1,4 +1,5 @@
 import asyncio
+import time
 
 from inu import const
 from inu.app import InuApp
@@ -52,23 +53,18 @@ class SwitchApp(InuApp):
             if new_state:
                 active += 1
 
+                if self.inu.settings.refire_delay and (sw.get_active_time() >= (self.inu.settings.refire_delay / 1000)):
+                    # Send a re-fire trigger
+                    await self.fire(i, self.get_code_for_switch(i))
+                    sw.active_time = time.time()
+
             if last_state != new_state:
-                # State changes
+                # State changed
                 self.logger.info(f"Switch {i}: {last_state} -> {new_state}")
 
                 if new_state:
                     # If we're moving into an active state, gather the right code and fire -
-                    if i > 5:
-                        # There are only 6 override codes
-                        code = -1
-                    else:
-                        code = int(getattr(self.inu.settings, f"sw_{i}"))
-
-                    # If override code is -1, use the default code (trigger_code)
-                    if code == -1:
-                        code = int(self.inu.settings.trigger_code)
-
-                    await self.fire(i, code)
+                    await self.fire(i, self.get_code_for_switch(i))
 
         if active != self.active:
             # Update the device status if the number of active switches changed
@@ -78,6 +74,19 @@ class SwitchApp(InuApp):
                 await self.inu.activate(f"Active: {active}")
             else:
                 await self.inu.deactivate()
+
+    def get_code_for_switch(self, index: int):
+        if index > 5:
+            # There are only 6 override codes
+            code = -1
+        else:
+            code = int(getattr(self.inu.settings, f"sw_{index}"))
+
+        # If override code is -1, use the default code (trigger_code)
+        if code == -1:
+            code = int(self.inu.settings.trigger_code)
+
+        return code
 
     async def fire(self, index, code):
         self.logger.info(f"Switch {index} firing; code {code}")